fix(issuance): align regression tests and address review feedback
