+++ apps/sdp-web/src/app/dashboard/issuance/[tokenId]/asset-profile/asset-profile-save-bar.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { Loader2 } from "lucide-react";
-import { AnimatePresence, motion, useReducedMotion } from "motion/react";
+import { AnimatePresence, domAnimation, LazyMotion, m, useReducedMotion } from "motion/react";
 import { Button } from "@/components/ui/button";
 import { useTranslations } from "@/i18n/provider";
 
@@ -23,49 +23,56 @@ export function AssetProfileSaveBar({
   const reducedMotion = useReducedMotion();
 
   return (
-    <AnimatePresence initial={false}>
-      {dirty ? (
-        <motion.div
-          key="save-bar"
-          initial={{ opacity: 0, y: reducedMotion ? 0 : 8 }}
-          animate={{ opacity: 1, y: 0 }}
-          exit={{ opacity: 0, y: reducedMotion ? 0 : 8 }}
-          transition={{ duration: reducedMotion ? 0 : 0.16 }}
-          className="sticky bottom-4 z-20"
-        >
-          <div className="mx-auto flex max-w-xl items-center justify-between gap-4 rounded-2xl border border-border-default bg-surface-raised px-5 py-3">
-            <p className="text-sm text-secondary">
-              {errorCount > 0
-                ? t("DashboardIssuance.saveBar.unsavedWithErrors", {
-                    count: errorCount,
-                    suffix: errorCount === 1 ? "" : "s",
-                  })
-                : t("DashboardIssuance.saveBar.unsaved")}
-            </p>
-            <div className="flex items-center gap-2">
-              <Button
-                type="button"
-                variant="secondary"
-                size="sm"
-                onClick={onDiscard}
-                disabled={saving}
-              >
-                {t("DashboardIssuance.saveBar.discard")}
-              </Button>
-              <Button type="button" size="sm" onClick={onSave} disabled={saving || errorCount > 0}>
-                {saving ? (
-                  <div className="flex items-center gap-2">
-                    <Loader2 className="h-4 w-4 animate-spin" />
-                    {t("DashboardIssuance.saveBar.saving")}
-                  </div>
-                ) : (
-                  t("DashboardIssuance.saveBar.save")
-                )}
-              </Button>
+    <LazyMotion features={domAnimation}>
+      <AnimatePresence initial={false}>
+        {dirty ? (
+          <m.div
+            key="save-bar"
+            initial={{ opacity: 0, y: reducedMotion ? 0 : 8 }}
+            animate={{ opacity: 1, y: 0 }}
+            exit={{ opacity: 0, y: reducedMotion ? 0 : 8 }}
+            transition={{ duration: reducedMotion ? 0 : 0.16 }}
+            className="sticky bottom-4 z-20"
+          >
+            <div className="mx-auto flex max-w-xl items-center justify-between gap-4 rounded-2xl border border-border-default bg-surface-raised px-5 py-3">
+              <p className="text-sm text-secondary">
+                {errorCount > 0
+                  ? t("DashboardIssuance.saveBar.unsavedWithErrors", {
+                      count: errorCount,
+                      suffix: errorCount === 1 ? "" : "s",
+                    })
+                  : t("DashboardIssuance.saveBar.unsaved")}
+              </p>
+              <div className="flex items-center gap-2">
+                <Button
+                  type="button"
+                  variant="secondary"
+                  size="sm"
+                  onClick={onDiscard}
+                  disabled={saving}
+                >
+                  {t("DashboardIssuance.saveBar.discard")}
+                </Button>
+                <Button
+                  type="button"
+                  size="sm"
+                  onClick={onSave}
+                  disabled={saving || errorCount > 0}
+                >
+                  {saving ? (
+                    <div className="flex items-center gap-2">
+                      <Loader2 className="h-4 w-4 animate-spin" />
+                      {t("DashboardIssuance.saveBar.saving")}
+                    </div>
+                  ) : (
+                    t("DashboardIssuance.saveBar.save")
+                  )}
+                </Button>
+              </div>
             </div>
-          </div>
-        </motion.div>
-      ) : null}
-    </AnimatePresence>
+          </m.div>
+        ) : null}
+      </AnimatePresence>
+    </LazyMotion>
   );
 }
