+++ apps/sdp-web/src/app/dashboard/issuance/[tokenId]/token-action-admin-forms.tsx
@@ -17,7 +17,6 @@ import {
   UserCog,
 } from "lucide-react";
 import {
-  type ComponentProps,
   type Dispatch,
   type ReactNode,
   type SetStateAction,
@@ -36,6 +35,7 @@ import { getPageCount, getPageSummary } from "../pagination.utils";
 import { fetchTokenAllowlistLabels, fetchTokenAllowlistPage } from "./asset-profile/allowlist.data";
 import { TOKEN_ALLOWLIST_KEY, TOKEN_ALLOWLIST_LABELS_KEY } from "./asset-profile/allowlist-cache";
 import { TokenActionCard } from "./token-action-card";
+import { ActionField } from "./token-action-field";
 import { TokenDisabledActionTooltip } from "./token-disabled-action-tooltip";
 import type {
   AdminAction,
@@ -55,7 +55,6 @@ import {
 import { TokenSignerSelect } from "./token-signer-select";
 import { TokenValidationMessage } from "./token-validation-message";
 import { TokenWalletAddressField } from "./token-wallet-address-field";
-import { useInlineValidationMessage } from "./use-inline-validation-message";
 
 interface TokenActionAdminFormsProps {
   activeAction: AdminAction | null;
@@ -975,93 +974,6 @@ function ControlListEntries({
         isPending={isPending}
         onRemove={onRemove}
       />
-    </div>
-  );
-}
-
-function ActionField({
-  label,
-  value,
-  onChange,
-  type = "text",
-  required = false,
-  pattern,
-  title,
-  min,
-  step,
-  placeholder,
-  inputMode,
-  description,
-  error,
-}: {
-  label: string;
-  value: string;
-  onChange: (value: string) => void;
-  type?: ComponentProps<typeof Input>["type"];
-  required?: boolean;
-  pattern?: string;
-  title?: string;
-  min?: string;
-  step?: string;
-  placeholder?: string;
-  inputMode?: ComponentProps<typeof Input>["inputMode"];
-  description?: string;
-  error?: string | null;
-}) {
-  const fieldId = useId();
-  const errorId = useId();
-  const t = useTranslations();
-  const [invalidNumber, setInvalidNumber] = useState(false);
-  const { message: nativeError, onInvalid, revalidate } = useInlineValidationMessage(label);
-  const fieldError = invalidNumber
-    ? t("DashboardIssuance.management.numberRequired")
-    : (error ?? nativeError);
-  const hasError = Boolean(fieldError);
-
-  return (
-    <div className="space-y-2">
-      <label
-        htmlFor={fieldId}
-        className="block text-[12px] leading-5 font-medium tracking-[0.02em] text-secondary"
-      >
-        {label}
-        {required ? (
-          <span aria-hidden className="text-destructive">
-            {" "}
-            *
-          </span>
-        ) : null}
-      </label>
-      {description ? <p className="text-[13px] leading-5 text-secondary">{description}</p> : null}
-      <Input
-        id={fieldId}
-        type={type === "number" ? "text" : type}
-        value={value}
-        required={required}
-        pattern={type === "number" ? "([0-9]+([.][0-9]*)?|[.][0-9]+)" : pattern}
-        title={title}
-        min={min}
-        step={step}
-        placeholder={placeholder}
-        inputMode={type === "number" ? "decimal" : inputMode}
-        aria-invalid={hasError}
-        aria-describedby={hasError ? errorId : undefined}
-        onInvalid={onInvalid}
-        onChange={(event) => {
-          if (type === "number" && !/^\d*\.?\d*$/.test(event.currentTarget.value)) {
-            event.currentTarget.value = value;
-            event.currentTarget.setCustomValidity(t("DashboardIssuance.management.numberRequired"));
-            setInvalidNumber(true);
-            return;
-          }
-          event.currentTarget.setCustomValidity("");
-          setInvalidNumber(false);
-          onChange(event.currentTarget.value);
-          revalidate(event.currentTarget);
-        }}
-        className="h-11 rounded-[12px] border-border-default bg-surface-raised px-4 shadow-none"
-      />
-      <TokenValidationMessage id={errorId} message={fieldError} />
     </div>
   );
 }
